Fix how we disable the release workflow
The previous approach introduced an initial step of only {if:false},
which is an invalid workflow.

I could've fixed that specifically, and placed the if correctly on each
step (or on the overall job), but I thought was cleaner to make the
workflow not listen for any events to start.