git-ignore stack.yaml.lock files
These lock files help with reproducibility and can alert developers if
the exact code behind a given resolver or extra-dep has changed since
the last build. However, certain tooling does or does not keep them
updated and this fact has led to confusion that is not worth this
benefit, particularly in projects with a lot of separate resolver files.
Therefore, we're going to start phasing out keeping them committed.