Many updates
* git-ignore stack.yaml.lock files

These lock files help with reproducibility and can alert developers if
the exact code behind a given resolver or extra-dep has changed since
the last build. However, certain tooling does or does not keep them
updated and this fact has led to confusion that is not worth this
benefit, particularly in projects with a lot of separate resolver files.
Therefore, we're going to start phasing out keeping them committed.

* Rename stack-lts files to only specify major

That's all that matters, and we don't want filenames (and test matrix
names) to change if minor versions fluctuate.

As part of this we move, to a hyphen-less suffix since that's what some
tooling (stack-all) expects, and it doesn't negatively affect anything
else.

* Use new haskell-actions organization

* Configure actions-based Restyled

* Add .stack-all configuration

* Add stack configurations for lts 21 through 23

* Make stack.yaml a symlink to stack-lts23.yaml

* Update release workflow

- `haskell-tag-action` doesn't need an explicit token
- `stack-upload-action` is deprecated, just use `stack upload`

* Add typically-disabled 9.8 warnings

* Add lts20-like extra-deps to lts21

* Update extra-deps in lts21

* Remove unnecessary pragma

Newer hspec-discover seems to include an export list.

* Move from library,tests to src,test

Anecdotally, `app,src,test` is more common that
`executables,library,tests`. It's also what `stack new` creates (`cabal
init` creates no directories). No one on the team expressed a strong
objection[^1], so we'll start a soft migration (i.e. use it in new
stuff, change at our leisure).

[^1]: https://freckleinc.slack.com/archives/C459XJBGR/p1702402421106759